Renamed ColumnFactory to ColumnHelper and added index renderer
