+++ src/components/shared/ColumnHelper.js
@@ -1,0 +1,51 @@
+import * as React from "react"
+
+export const GENERATED_ID_COL_INDEX = 0
+export const RESTRICTED_COL_INDEX = 1
+
+export function indexRender(value, tableMeta, bookData, currentBook) {
+  let indices = value.split(",")
+  return (
+    <div>
+      {indices.map((index, count) => {
+        let idAndPage = index.split(":").map(s => s.trim())
+        let book = bookData.allBooksYaml.edges
+          .map(({ node }) => node)
+          .filter(node => node.generatedId === idAndPage[0])
+        console.log(book)
+        return currentBook !== idAndPage[0] ? (
+          <span key={tableMeta.rowData[GENERATED_ID_COL_INDEX]}>
+            <Link to={`/books/${idAndPage[0]}/`}>{book[0].name}</Link>:
+            {idAndPage[1]}
+            {count !== indices.length - 1 ? ", " : ""}
+          </span>
+        ) : (
+          <span key={tableMeta.rowData[GENERATED_ID_COL_INDEX]}>
+            {book[0].name}:{idAndPage[1]}
+            {count !== indices.length - 1 ? ", " : ""}
+          </span>
+        )
+      })}
+    </div>
+  )
+}
+
+export const makeColumns = (columns, includeRestricted) => {
+  includeRestricted = includeRestricted || false
+
+  return [
+    {
+      name: "generatedId",
+      options: { display: false, viewColumns: false, filter: false },
+    },
+    ...(includeRestricted
+      ? [
+          {
+            name: "restricted",
+            options: { display: false, viewColumns: false, filter: false },
+          },
+        ]
+      : []),
+    ...columns,
+  ]
+}
